vm/audio-pipeline/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Audio I/O
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1

# Signal processing
numpy==1.26.4
scipy==1.13.1

# Noise reduction — spectral subtraction
noisereduce==3.0.2

# Transcription — Whisper large-v3 (CPU int8)
faster-whisper==1.0.3

# Speaker diarization (optional — requires HF_TOKEN)
pyannote.audio==3.1.1

# Loudness normalisation
pyloudnorm==0.1.1

# Visualisation
matplotlib==3.9.0
Pillow==10.3.0

# Utilities
httpx==0.27.0

# Neural noise reduction (DeepFilterNet v3)
DeepFilterNet==0.5.6

# System monitoring
psutil==5.9.8
